Move red tech 0 units to their own file
