+++ src/data/units.ts
@@ -1,10 +1,8 @@
 import { isPatrolling } from "../framework/accessors";
-import { ResolvableEffect } from "../framework/types";
 
 import { REQUIRE_ALL_CARD_PROPERTIES } from "./config";
 import {
   active,
-  constantModifiers,
   constantParam,
   effectBase,
   getProperties,
@@ -12,6 +10,7 @@ import {
   trigger,
   valueParam,
 } from "./helpers";
+import { RedTechZero } from "./specs/red-0";
 
 const unitBoostCosts: Record<string, number | undefined> = {
   // such empty
@@ -27,80 +26,11 @@ export const getUnitProperties = getProperties(id => {
   let aid = 1;
   switch (id) {
   case "Bloodrage Ogre":
-    return {
-      triggeredAbilities: [
-        trigger("END_OF_TURN", ($, I) => {
-          if ($.state.activePlayer !== I.controller) {
-            return [];
-          }
-          if (I.arrivalFatigue) {
-            return [];
-          }
-          if (I.memory.attackedThisTurn === "true") {
-            return [];
-          }
-          return [
-            {
-              ...effectBase(id, I, "BOUNCE_TO_HAND"),
-              target: valueParam("INSTANCE", I.id),
-            },
-          ];
-        }),
-        trigger("THIS_ATTACKS", (_, I) => {
-          I.memory.attackedThisTurn = "true";
-          return [];
-        }),
-        trigger("UPKEEP", (_, I) => {
-          I.memory.attackedThisTurn = "false";
-          return [];
-        }),
-      ],
-    };
+    return RedTechZero.BloodrageOgre(id);
   case "Bombaster":
-    return {
-      activatedAbilities: [
-        active(id, aid++, [
-          { type: "GOLD", amount: 1 },
-          { type: "SACRIFICE_THIS" },
-        ], (_, I) => [
-          {
-            ...effectBase(id, I, "DAMAGE"),
-            target: queryParam("INSTANCE", {
-              patrolling: true,
-              type: "UNIT",
-            }),
-            amount: constantParam(2),
-          },
-        ]),
-      ],
-    };
+    return RedTechZero.Bombaster(id);
   case "Careless Musketeer":
-    return {
-      activatedAbilities: [
-        active(id, aid++, [ { type: "EXHAUST_THIS" } ], ($, I) => {
-          const effects: Array<ResolvableEffect> = [
-            {
-              ...effectBase(id, I, "DAMAGE"),
-              target: queryParam("INSTANCE", {
-                type: [ "BUILDING", "UNIT" ],
-              }),
-              amount: constantParam(1),
-            },
-          ];
-
-          const base = $.getPlayer(I.controller)?.base;
-          if (base != null) {
-            effects.push({
-              ...effectBase(id, I, "DAMAGE"),
-              target: valueParam("INSTANCE", base),
-              amount: constantParam(1),
-            });
-          }
-
-          return effects;
-        }),
-      ],
-    };
+    return RedTechZero.CarelessMusketeer(id);
   case "Ironbark Treant":
     return {
       continuousModifiers: [
@@ -129,38 +59,9 @@ export const getUnitProperties = getProperties(id => {
       ],
     };
   case "Mad Man":
-    return {
-      traits: [ "HASTE" ],
-    };
+    return RedTechZero.MadMan(id);
   case "Makeshift Rambaster":
-    return {
-      traits: [ "HASTE", "NO_PATROL" ],
-      triggeredAbilities: [
-        trigger("THIS_ATTACKS", ($, I, e) => {
-          const targetCard = $.data.lookupCard(e.instance.card);
-          if (targetCard.type !== "BUILDING") {
-            return [];
-          }
-          return [
-            {
-              ...effectBase(id, I, "MODIFY"),
-              target: valueParam("INSTANCE", I.id),
-              modifiers: constantModifiers([
-                {
-                  expiration: "END_OF_COMBAT",
-                  sourceCard: I.card,
-                  effect: {
-                    type: "ATTRIBUTE",
-                    attribute: "ATTACK",
-                    amount: 2,
-                  },
-                },
-              ]),
-            },
-          ];
-        }),
-      ],
-    };
+    return RedTechZero.MakeshiftRambaster(id);
   case "Merfolk Prospector":
     return {
       activatedAbilities: [
@@ -181,9 +82,7 @@ export const getUnitProperties = getProperties(id => {
       ],
     };
   case "Nautical Dog":
-    return {
-      attributes: { FRENZY: 1 },
-    };
+    return RedTechZero.NauticalDog(id);
   case "Playful Panda":
     return {
       triggeredAbilities: [
